migration: update namespaces
