+++ Kontent.Ai.AspNetCore.Tests/.gitattributes
