+++ Kontent.Ai.AspNetCore.Tests/Kontent.Ai.AspNetCore.Tests.csproj
@@ -25,7 +25,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Kentico.Kontent.AspNetCore\Kentico.Kontent.AspNetCore.csproj" />
+    <ProjectReference Include="..\Kontent.Ai.AspNetCore\Kontent.Ai.AspNetCore.csproj" />
   </ItemGroup>
 
   <ItemGroup>
