+++ Kontent.Ai.AspNetCore.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30011.22
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kentico.Kontent.AspNetCore", "Kentico.Kontent.AspNetCore\Kentico.Kontent.AspNetCore.csproj", "{7C996EEF-91B2-4DD3-9550-FA978667D7EF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kontent.Ai.AspNetCore", "Kontent.Ai.AspNetCore\Kontent.Ai.AspNetCore.csproj", "{7C996EEF-91B2-4DD3-9550-FA978667D7EF}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kentico.Kontent.AspNetCore.Tests", "Kentico.Kontent.AspNetCore.Tests\Kentico.Kontent.AspNetCore.Tests.csproj", "{3426DEB9-0D82-44BE-AE88-A0CD9D10E1A3}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kontent.Ai.AspNetCore.Tests", "Kontent.Ai.AspNetCore.Tests\Kontent.Ai.AspNetCore.Tests.csproj", "{3426DEB9-0D82-44BE-AE88-A0CD9D10E1A3}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{C694E5C3-563D-4C3F-91A4-A4B004D95B44}"
 	ProjectSection(SolutionItems) = preProject
