+++ Kontent.Ai.AspNetCore/Webhooks/Models/Common.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
-namespace Kentico.Kontent.AspNetCore.Webhooks.Models
+namespace Kontent.Ai.AspNetCore.Webhooks.Models
 {
     /// <summary>
     /// The Message object contains information about the origin of the notification.
