+++ Kontent.Ai.AspNetCore/Webhooks/Models/Reference.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
-namespace Kentico.Kontent.AspNetCore.Webhooks.Models
+namespace Kontent.Ai.AspNetCore.Webhooks.Models
 {
     /// <summary>
     /// A reference object
