+++ Kontent.Ai.AspNetCore/Webhooks/WebhookOptions.cs
@@ -1,4 +1,4 @@
-﻿namespace Kentico.Kontent.AspNetCore.Webhooks
+﻿namespace Kontent.Ai.AspNetCore.Webhooks
 {
     /// <summary>
     /// A configuration object that allows to adjust the Kentico Kontent webhook behavior.
