+++ README.md
@@ -1,5 +1,5 @@
-[![NuGet](https://img.shields.io/nuget/vpre/Kentico.Kontent.AspNetCore.svg)](https://www.nuget.org/packages/Kentico.Kontent.AspNetCore/)
-[![Downloads](https://img.shields.io/nuget/dt/Kentico.Kontent.AspNetCore.svg)](https://www.nuget.org/packages/Kentico.Kontent.AspNetCore/)
+[![NuGet](https://img.shields.io/nuget/vpre/Kontent.Ai.AspNetCore.svg)](https://www.nuget.org/packages/Kontent.Ai.AspNetCore/)
+[![Downloads](https://img.shields.io/nuget/dt/Kontent.Ai.AspNetCore.svg)](https://www.nuget.org/packages/Kontent.Ai.AspNetCore/)
 [![Build & Test](https://github.com/Kentico/kontent-aspnetcore/actions/workflows/integrate.yml/badge.svg)](https://github.com/Kentico/kontent-aspnetcore/actions/workflows/integrate.yml)
 [![codecov](https://codecov.io/gh/Kentico/kontent-aspnetcore/branch/master/graph/badge.svg?token=U4Y9PQDW6Q)](https://codecov.io/gh/Kentico/kontent-aspnetcore)
 [![Stack Overflow](https://img.shields.io/badge/Stack%20Overflow-ASK%20NOW-FE7A16.svg?logo=stackoverflow&logoColor=white)](https://stackoverflow.com/tags/kentico-kontent)
@@ -22,7 +22,7 @@ Useful for rendering responsive images. Supports Assets and Inline images in ric
 
 **_ViewImports.cshtml**
 ```razor
-@addTagHelper *, Kentico.Kontent.AspNetCore
+@addTagHelper *, Kontent.Ai.AspNetCore
 ```
 
 **Startup.cs**
